# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "lumino 记账本服务",
        "title": "Lumino",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/api/v1/user": {
            "get": {
                "tags": [
                    "用户"
                ],
                "summary": "获取用户信息",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "用户id",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "用户结果",
                        "schema": {
                            "$ref": "#/definitions/model.User"
                        }
                    },
                    "400": {
                        "description": "请求体异常",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "服务端异常",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "用户"
                ],
                "summary": "修改用户信息",
                "parameters": [
                    {
                        "type": "number",
                        "description": "余额",
                        "name": "balance",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "默认账本id",
                        "name": "default_account_book_id",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "integer"
                        },
                        "collectionFormat": "csv",
                        "description": "朋友列表",
                        "name": "friend",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "账号名称，昵称，全局唯一",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "object",
                        "description": "余额详情",
                        "name": "balance_detail",
                        "in": "query"
                    },
                    {
                        "type": "file",
                        "description": "用户头像",
                        "name": "icon_file",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "用户修改后结果",
                        "schema": {
                            "$ref": "#/definitions/model.User"
                        }
                    },
                    "400": {
                        "description": "请求体异常",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "服务端异常",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "用户"
                ],
                "summary": "注册用户信息",
                "parameters": [
                    {
                        "type": "number",
                        "description": "余额",
                        "name": "balance",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "默认账本id",
                        "name": "default_account_book_id",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "integer"
                        },
                        "collectionFormat": "csv",
                        "description": "朋友列表",
                        "name": "friend",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "账号名称，昵称，全局唯一",
                        "name": "name",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "object",
                        "description": "余额详情",
                        "name": "balance_detail",
                        "in": "query"
                    },
                    {
                        "type": "file",
                        "description": "用户头像",
                        "name": "icon_file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "注册结果",
                        "schema": {
                            "$ref": "#/definitions/model.User"
                        }
                    },
                    "400": {
                        "description": "请求体异常",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "服务端异常",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "用户"
                ],
                "summary": "删除用户信息",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "用户id",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "请求体异常",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "服务端异常",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "model.User": {
            "type": "object",
            "properties": {
                "balance": {
                    "description": "余额",
                    "type": "number"
                },
                "balance_detail": {
                    "description": "BalanceDetail 余额详情\n@swagger:type object\n@additionalProperties type=number format=double\n@example {\"temperature\":36.5,\"humidity\":0.42}",
                    "type": "object",
                    "additionalProperties": {
                        "type": "number"
                    }
                },
                "created_at": {
                    "description": "创建时间",
                    "type": "string"
                },
                "default_account_book_id": {
                    "description": "默认账本id",
                    "type": "integer"
                },
                "deleted_at": {
                    "description": "DeletedAt 删除时间",
                    "type": "string",
                    "example": "2025-03-26T00:00:00Z"
                },
                "friend": {
                    "description": "朋友列表",
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "icon_url": {
                    "description": "用户头像的对象存储地址",
                    "type": "string"
                },
                "id": {
                    "description": "主键id",
                    "type": "integer"
                },
                "name": {
                    "description": "账号名称，昵称，全局唯一",
                    "type": "string"
                },
                "updated_at": {
                    "description": "更新时间",
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}